pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agent-harness-skills"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk",
    "anyio",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "anthropic>=0.52.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "playwright",
    "pytest-playwright",
    "jsonschema>=4.18",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0",
    "pre-commit>=3.7",
    "pytest-cov>=4.0",
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
]
dashboard = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "rich>=13.0.0",
    "click>=8.1.0",
]
languages = [
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-java>=0.23.0",
]

[project.entry-points."harness_skills.analyzers"]
python = "harness_skills.analyzers.python_analyzer:PythonAnalyzer"
typescript = "harness_skills.analyzers.typescript_analyzer:TypeScriptAnalyzer"
go = "harness_skills.analyzers.go_analyzer:GoAnalyzer"
rust = "harness_skills.analyzers.rust_analyzer:RustAnalyzer"
java = "harness_skills.analyzers.java_analyzer:JavaAnalyzer"
csharp = "harness_skills.analyzers.csharp_analyzer:CSharpAnalyzer"

[project.entry-points."harness_skills.gates"]
# Third-party gates register here

[project.entry-points."harness_skills.generators"]
# Third-party generators register here

[tool.setuptools.packages.find]
where = ["."]
include = ["harness_skills*", "harness_tools*", "harness_dashboard", "log_format_linter", "dom_snapshot_utility"]

# ---------------------------------------------------------------------------
# Ruff — linter + formatter
# https://docs.astral.sh/ruff/
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100

# Directories ruff should never touch
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
    "*.egg-info",
    "dist",
    "build",
    "observability",          # third-party grafana/loki config dirs
]

[tool.ruff.lint]
# Rule sets enabled:
#   E/W  – pycodestyle errors & warnings
#   F    – Pyflakes (undefined names, unused imports, …)
#   I    – isort (import ordering)
#   N    – pep8-naming  ← enforces class/function/variable naming conventions
#   UP   – pyupgrade    ← modernise syntax for Python 3.12
#   B    – flake8-bugbear
#   SIM  – flake8-simplify
#   PTH  – flake8-use-pathlib
select = ["E", "W", "F", "I", "N", "UP", "B", "SIM", "PTH"]

ignore = [
    # Allow long lines when they contain a URL
    "E501",
    # `assert` is fine in tests
    "B011",
    # Allow `*` imports in __init__ re-export files
    "F401",
]

# Per-file rule overrides
[tool.ruff.lint.per-file-ignores]
# Test files: relax naming rules (fixture functions, parametrize ids, etc.)
"tests/**/*.py"     = ["N802", "N803", "N806"]
"test_*.py"         = ["N802", "N803", "N806"]
# Example files are allowed to be more permissive
"*_example.py"      = ["N802", "E402"]
# CLI entry-points that shadow builtins
"harness_skills/cli/*.py" = ["N802"]

[tool.ruff.lint.pep8-naming]
# Allow these patterns as "class-like" names that may use PascalCase
classmethod-decorators = ["classmethod", "pydantic.validator", "pydantic.model_validator"]

[tool.ruff.lint.isort]
known-first-party = [
    "harness_skills",
    "harness_tools",
    "harness_dashboard",
    "log_format_linter",
    "dom_snapshot_utility",
]
force-sort-within-sections = true

[tool.ruff.format]
# Use double quotes (PEP 8 default)
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# --browser chromium  targets Chromium for the tests/browser/ suite.
# pytest-playwright picks this up automatically; non-browser tests ignore it.
addopts = "-v --tb=short --browser chromium"
